Add files via upload
This dataset is originally from the National Institute of Diabetes and Digestive 
and Kidney Diseases. The objective of the dataset is to diagnostically predict 
whether a patient has diabetes based on certain diagnostic measurements 
included in the dataset. Several constraints were placed on the selection of 
these instances from a larger database. In particular, all patients here are 
females at least 21 years old of Pima Indian heritage